requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cloudinary==1.26.0
cycler==0.10.0
dlib==19.22.0
dnspython==2.1.0
easy-ocr==0.1.1
easyocr==1.4
email-validator==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mongoengine==0.23.1
networkx==2.6.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.1
pyparsing==2.4.7
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.1
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tifffile==2021.7.30
torch==1.5.1+cpu
torchvision==0.6.1+cpu
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
